moved pages to another folder | issue #28
