+++ src/pages/Cart/index.ts
@@ -1,0 +1,3 @@
+import CartPage from './CartPage';
+
+export { CartPage };
