+++ src/pages/Home/ChoosePick/ChoosePick.styled.ts
