+++ src/pages/Home/ChoosePick/index.ts
