+++ src/pages/Home/FoodCategory/index.ts
