+++ src/pages/Menu/index.ts
@@ -1,0 +1,3 @@
+import MenuPage from './MenuPage';
+
+export { MenuPage };
