+++ src/pages/index.ts
@@ -1,0 +1,5 @@
+import { CartPage } from './Cart';
+import { HomePage } from './Home';
+import { MenuPage } from './Menu';
+
+export { CartPage, HomePage, MenuPage };
