+++ src/sections/index.ts
@@ -1,11 +1,2 @@
-export { AboutUs } from './AboutUs';
-export { Achievement } from './Achievement';
-export { ActiveProcess } from './ActiveProcess';
-export { ChoosePick } from './ChoosePick';
-export { FoodCategory } from './FoodCategory';
 export { Footer } from './Footer';
 export { Header, NavBar } from './Header';
-export { LatesBlogPost } from './LatesBlogPost';
-export { OurChef } from './OurChef';
-export { Testimonials } from './Testimonials';
-export { WhyChooseUs } from './WhyChooseUs';
