--- src/views/index.ts
@@ -1,5 +1,0 @@
-import CartPage from './CartPage';
-import HomePage from './HomePage';
-import MenuPage from './MenuPage';
-
-export { CartPage, HomePage, MenuPage };
